*squash* *squash* *squash*
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@1844 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ DarkMod/AI/States/State.cpp
@@ -1142,7 +1142,7 @@ void State::OnAICommMessage(CAIComm_Message* message)
 			{
 				idAI* issuer = static_cast<idAI*>(issuingEntity);
 				// Note: We deliberately don't care if the issuer is a friend or not
-				float warningAmount = issuer->GetMind()->GetMemory().countEvidenceOfIntruders;
+				int warningAmount = issuer->GetMind()->GetMemory().countEvidenceOfIntruders;
 				
 				if (memory.countEvidenceOfIntruders < warningAmount)
 				{
